refactor(xml): extract Tag & TagKind
