+++ pdf_craft/xml/parser.py
@@ -1,7 +1,7 @@
 from io import StringIO
-from dataclasses import dataclass
 from typing import Generator, Iterable
 from enum import auto, Enum
+from .tag import Tag, TagKind
 
 
 _SPACES = (" ", "\n")
@@ -22,40 +22,6 @@ class _ParsedResult(Enum):
   Continue = auto()
   Success = auto()
   Failed = auto()
-
-class TagKind(Enum):
-  OPENING = auto()
-  CLOSING = auto()
-  SELF_CLOSING = auto()
-
-@dataclass
-class Tag:
-  kind: TagKind
-  name: str
-  proto: str
-  attributes: list[tuple[str, str]]
-
-  def __str__(self):
-    buffer = StringIO()
-    buffer.write("<")
-    if self.kind == TagKind.CLOSING:
-      buffer.write("/")
-    buffer.write(self.name)
-    if len(self.attributes) > 0:
-      buffer.write(" ")
-      for i, (attr_name, attr_value) in enumerate(self.attributes):
-        buffer.write(attr_name)
-        buffer.write("=")
-        buffer.write("\"")
-        buffer.write(attr_value)
-        buffer.write("\"")
-        if i < len(self.attributes) - 1:
-          buffer.write(" ")
-    if self.kind == TagKind.SELF_CLOSING:
-      buffer.write("/>")
-    else:
-      buffer.write(">")
-    return buffer.getvalue()
 
 def parse_tags(chars: Iterable[str]) -> Generator[str | Tag, None, None]:
   yield from _XMLTagsParser().do(chars)
@@ -176,7 +142,7 @@ class _XMLTagsParser:
   def _generate_by_result(self, parsed_result: _ParsedResult) -> Generator[str | Tag, None, None]:
     if parsed_result == _ParsedResult.Success:
       assert self._tag is not None
-      if self._is_tag_valid(self._tag):
+      if self._tag.is_valid():
         outside_text = self._outside_buffer.getvalue()
         self._clear_buffer(self._outside_buffer)
         self._clear_buffer(self._tag_buffer)
@@ -194,28 +160,6 @@ class _XMLTagsParser:
       self._outside_buffer.write(self._tag_buffer.getvalue())
       self._clear_buffer(self._tag_buffer)
       self._phase = _Phase.OUTSIDE
-
-  def _is_tag_valid(self, tag: Tag) -> bool:
-    if tag.kind == TagKind.CLOSING and len(tag.attributes) > 0:
-      return False
-
-    for name in self._iter_tag_names(tag):
-      # https://www.w3schools.com/xml/xml_elements.asp
-      if name == "":
-        return False
-      char = name[0]
-      if char == "_":
-        continue
-      if "a" <= char <= "z" or "A" <= char <= "Z":
-        continue
-      return False
-
-    return True
-
-  def _iter_tag_names(self, tag: Tag):
-    yield tag.name
-    for attr_name, _ in tag.attributes:
-      yield attr_name
 
   def _clear_buffer(self, buffer: StringIO):
     buffer.truncate(0)
